backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose
passlib[bcrypt]
python-multipart
python-dotenv
